Fix floordiv for i32 x/y inputs

+++ src/runtime/lpython_builtin.py
@@ -481,7 +481,7 @@ def _lpython_floordiv(a: f32, b: f32) -> f32:
 
 @overload
 def _lpython_floordiv(a: i32, b: i32) -> i32:
-    r: f32
+    r: f64
     r = a/b
     result: i32
     result = int(r)
